Test: testing modules ,
Gradle modifications were made. There are still some exceptions to evaluate and interfaces. There are variables declared package therefore not known whether they are public or private access . There are other variables are difficult to handle, because the core fermat belong .
