--- TKY/plugin/sub_app_module/fermat-tky-plugin-sub-app-module-artist-identity-bitdubai/src/test/java/com/bitdubai/fermat_tky_plugin/layer/sub_app_module/artist_identity/developer/bitdubai/revision1/structure/ArtistIdentityManagerTest/testcreateArtistIdentity.java
@@ -1,73 +1,0 @@
-package com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.revision1.structure.ArtistIdentityManagerTest;
-
-import com.bitdubai.fermat_api.layer.all_definition.common.system.utils.AddonVersionReference;
-import com.bitdubai.fermat_api.layer.all_definition.common.system.utils.PluginVersionReference;
-import com.bitdubai.fermat_api.layer.all_definition.enums.Addons;
-import com.bitdubai.fermat_api.layer.all_definition.enums.PlatformComponents;
-import com.bitdubai.fermat_api.layer.all_definition.enums.Plugins;
-import com.bitdubai.fermat_api.layer.all_definition.enums.UISource;
-import com.bitdubai.fermat_api.layer.all_definition.events.interfaces.FermatEvent;
-import com.bitdubai.fermat_api.layer.all_definition.navigation_structure.enums.Wallets;
-import com.bitdubai.fermat_api.layer.dmp_engine.sub_app_runtime.enums.SubApps;
-import com.bitdubai.fermat_pip_api.layer.platform_service.error_manager.enums.UnexpectedAddonsExceptionSeverity;
-import com.bitdubai.fermat_pip_api.layer.platform_service.error_manager.enums.UnexpectedPlatformExceptionSeverity;
-import com.bitdubai.fermat_pip_api.layer.platform_service.error_manager.enums.UnexpectedPluginExceptionSeverity;
-import com.bitdubai.fermat_pip_api.layer.platform_service.error_manager.enums.UnexpectedSubAppExceptionSeverity;
-import com.bitdubai.fermat_pip_api.layer.platform_service.error_manager.enums.UnexpectedUIExceptionSeverity;
-import com.bitdubai.fermat_pip_api.layer.platform_service.error_manager.enums.UnexpectedWalletExceptionSeverity;
-import com.bitdubai.fermat_pip_api.layer.platform_service.error_manager.interfaces.ErrorManager;
-import com.bitdubai.fermat_tky_api.all_definitions.enums.ArtistAcceptConnectionsType;
-import com.bitdubai.fermat_tky_api.all_definitions.enums.ExposureLevel;
-import com.bitdubai.fermat_tky_api.all_definitions.enums.ExternalPlatform;
-import com.bitdubai.fermat_tky_api.all_definitions.exceptions.IdentityNotFoundException;
-import com.bitdubai.fermat_tky_api.layer.identity.artist.exceptions.ArtistIdentityAlreadyExistsException;
-import com.bitdubai.fermat_tky_api.layer.identity.artist.exceptions.CantCreateArtistIdentityException;
-import com.bitdubai.fermat_tky_api.layer.identity.artist.exceptions.CantGetArtistIdentityException;
-import com.bitdubai.fermat_tky_api.layer.identity.artist.exceptions.CantListArtistIdentitiesException;
-import com.bitdubai.fermat_tky_api.layer.identity.artist.exceptions.CantUpdateArtistIdentityException;
-import com.bitdubai.fermat_tky_api.layer.identity.artist.interfaces.Artist;
-import com.bitdubai.fermat_tky_api.layer.identity.artist.interfaces.TokenlyArtistIdentityManager;
-import com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.version_1.structure.ArtistIdentityManager;
-
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.ExpectedException;
-
-import java.util.List;
-import java.util.UUID;
-
-/**
- * Created by MACUARE on 19/04/16.
- */
-public class testcreateArtistIdentity {
-
-    private ArtistIdentityManager artistIdentityManager;
-
-
-    @Rule
-    public ExpectedException exception = ExpectedException.none();
-
-    @Before
-    public void setUp() throws Exception {
-        Assert.assertNull(artistIdentityManager);
-        artistIdentityManager = new ArtistIdentityManager(null, null);
-        Assert.assertNotNull(artistIdentityManager);
-    }
-
-    @Test
-    public void testcreateArtistIdentity() throws ArtistIdentityAlreadyExistsException, CantCreateArtistIdentityException {
-       exception.expect(NullPointerException.class);
-       Artist artist =  artistIdentityManager.createArtistIdentity(
-                "jon doe",
-                "image".getBytes(),
-                "123456",
-                ExternalPlatform.DEFAULT_EXTERNAL_PLATFORM,
-                ExposureLevel.DEFAULT_EXPOSURE_LEVEL,
-                ArtistAcceptConnectionsType.DEFAULT_ARTIST_ACCEPT_CONNECTION_TYPE);
-
-       Assert.assertNull(artist);
-
-    }
-}
