+++ TKY/plugin/sub_app_module/fermat-tky-plugin-sub-app-module-artist-identity-bitdubai/src/test/java/com/bitdubai/fermat_tky_plugin/layer/sub_app_module/artist_identity/developer/bitdubai/version_1/ArtistIdentityPluginRootTest/TestgetFanIdentityManager.java
@@ -1,4 +1,4 @@
-package com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.revision1.ArtistIdentityPluginRootTest;
+package com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.version_1.ArtistIdentityPluginRootTest;
 
 import com.bitdubai.fermat_tky_api.all_definitions.exceptions.TKYException;
 import com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.version_1.ArtistIdentityPluginRoot;
