+++ TKY/plugin/sub_app_module/fermat-tky-plugin-sub-app-module-artist-identity-bitdubai/src/test/java/com/bitdubai/fermat_tky_plugin/layer/sub_app_module/artist_identity/developer/bitdubai/version_1/DeveloperBitDubaiTest/TestgetAddress.java
@@ -1,4 +1,4 @@
-package com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.revision1.DeveloperBitDubaiTest;
+package com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.version_1.DeveloperBitDubaiTest;
 
 import com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.DeveloperBitDubai;
 
@@ -15,6 +15,7 @@ public class TestgetAddress {
 
     @Before
     public void setUp() throws Exception {
+        System.out.println("initializing");
         Assert.assertNull(developerBitDubai);
         developerBitDubai = new DeveloperBitDubai();
         Assert.assertNotNull(developerBitDubai);
@@ -22,6 +23,7 @@ public class TestgetAddress {
 
     @Test
     public void testgetAddress() throws Exception {
+        System.out.println("testgetAddress");
         String actual = developerBitDubai.getAddress();
         String expected = "19qRypu7wrndwW4FRCxU1JPr5we134Q3eh";
         Assert.assertTrue(actual.length() >= 34);
