+++ TKY/plugin/sub_app_module/fermat-tky-plugin-sub-app-module-artist-identity-bitdubai/src/test/java/com/bitdubai/fermat_tky_plugin/layer/sub_app_module/artist_identity/developer/bitdubai/version_1/structure/ArtistIdentityManagerTest/testcreateArtistIdentity.java
@@ -1,0 +1,51 @@
+package com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.version_1.structure.ArtistIdentityManagerTest;
+
+import com.bitdubai.fermat_tky_api.all_definitions.enums.ArtistAcceptConnectionsType;
+import com.bitdubai.fermat_tky_api.all_definitions.enums.ExposureLevel;
+import com.bitdubai.fermat_tky_api.all_definitions.enums.ExternalPlatform;
+import com.bitdubai.fermat_tky_api.layer.identity.artist.exceptions.ArtistIdentityAlreadyExistsException;
+import com.bitdubai.fermat_tky_api.layer.identity.artist.exceptions.CantCreateArtistIdentityException;
+import com.bitdubai.fermat_tky_api.layer.identity.artist.interfaces.Artist;
+import com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.version_1.structure.ArtistIdentityManager;
+
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.ExpectedException;
+
+/**
+ * Created by MACUARE on 19/04/16.
+ */
+public class testcreateArtistIdentity {
+
+    private ArtistIdentityManager artistIdentityManager;
+
+
+    @Rule
+    public ExpectedException exception = ExpectedException.none();
+
+    @Before
+    public void setUp() throws Exception {
+        System.out.println("initializing");
+        Assert.assertNull(artistIdentityManager);
+        artistIdentityManager = new ArtistIdentityManager(null, null);
+        Assert.assertNotNull(artistIdentityManager);
+    }
+
+    @Test
+    public void testcreateArtistIdentity() throws ArtistIdentityAlreadyExistsException, CantCreateArtistIdentityException {
+        System.out.println("testcreateArtistIdentity");
+        exception.expect(NullPointerException.class);
+        Artist artist =  artistIdentityManager.createArtistIdentity(
+                "jon doe",
+                "image".getBytes(),
+                "123456",
+                ExternalPlatform.DEFAULT_EXTERNAL_PLATFORM,
+                ExposureLevel.DEFAULT_EXPOSURE_LEVEL,
+                ArtistAcceptConnectionsType.DEFAULT_ARTIST_ACCEPT_CONNECTION_TYPE);
+
+        Assert.assertNull(artist);
+
+    }
+}
