+++ TKY/plugin/sub_app_module/fermat-tky-plugin-sub-app-module-artist-identity-bitdubai/src/test/java/com/bitdubai/fermat_tky_plugin/layer/sub_app_module/artist_identity/developer/bitdubai/version_1/structure/ArtistIdentityManagerTest/testlistIdentitiesFromCurrentDeviceUser.java
@@ -1,6 +1,5 @@
-package com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.revision1.structure.ArtistIdentityManagerTest;
+package com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.version_1.structure.ArtistIdentityManagerTest;
 
-import com.bitdubai.fermat_pip_api.layer.platform_service.error_manager.interfaces.ErrorManager;
 import com.bitdubai.fermat_tky_api.layer.identity.artist.exceptions.CantListArtistIdentitiesException;
 import com.bitdubai.fermat_tky_api.layer.identity.artist.interfaces.Artist;
 import com.bitdubai.fermat_tky_plugin.layer.sub_app_module.artist_identity.developer.bitdubai.version_1.structure.ArtistIdentityManager;
@@ -24,6 +23,7 @@ public class testlistIdentitiesFromCurrentDeviceUser {
 
     @Before
     public void setUp() throws Exception {
+        System.out.println("initializing");
         Assert.assertNull(artistIdentityManager);
         artistIdentityManager = new ArtistIdentityManager(null,null);
         Assert.assertNotNull(artistIdentityManager);
@@ -32,6 +32,7 @@ public class testlistIdentitiesFromCurrentDeviceUser {
 
     @Test
     public void testlistIdentitiesFromCurrentDeviceUser() throws CantListArtistIdentitiesException {
+        System.out.println("testlistIdentitiesFromCurrentDeviceUser");
         exception.expect(NullPointerException.class);
 
         List<Artist> artists = artistIdentityManager.listIdentitiesFromCurrentDeviceUser();
